Extracts a HasAttachments trait and use that in the comments model
